+++ src/onSnapshot/onSnapshot.test.ts
@@ -61,9 +61,6 @@ describe('test onSnapshot', () => {
 				},
 				() => {
 					//
-				},
-				() => {
-					//
 				}
 			)
 		})